requirements.txt
Django==1.8.13
django-ckeditor==5.0.3
django-haystack==2.4.1
django-taggit==0.20.2
jwt==0.3.2
Markdown==2.6.6
Pillow==3.3.0
pycrypto==2.6.1
pysolr==3.3.2
pytz==2016.6.1
